[MOD] csv and single account creation popup content put in their own component, updated report creation popup
