+++ src/css/Components/Popup/popup.scss
@@ -98,13 +98,16 @@
         gap: 10px;
 
         input, select, textarea {
-            width: 350px;
+            width: 90%;
             border: 2px solid #4f23e2;
             border-radius: 15px;
             height: 25px;
             padding-left: 10px;
             background-color: white;
+        }
 
+        textarea {
+            height: 15%;
         }
 
         #select-classes .css-13cymwt-control, #select-classes .react-select__menu {
